Name and whitespace fixes to the test app
